refactor code to accomodate bookmark functionality
