+++ src/components/ui/profileList/ProfileList.js
@@ -1,37 +1,15 @@
-import React, { useState, useEffect } from 'react';
-import axios from 'axios';
+import React from 'react';
+import PropTypes from 'prop-types';
 
 import PersonPreview from '../profilePreview/ProfilePreview';
-import StyledProfileList from './styles';
-import Spinner from '../../common/Spinner';
-import config from '../../../utils/config';
+import { StyledProfileList, H2 } from './styles';
 import Container from '../../common/Container';
 
-const { apiBaseUrl } = config;
-
-const ProfileList = () => {
-  const [profiles, setProfiles] = useState([]); // this will hold the profles list fetched from the API
-  const [loading, setLoading] = useState(true);
-
-  useEffect(() => {
-    const fetchdata = async () => {
-      try {
-        const response = await axios.get(`${apiBaseUrl}/people`);
-        setProfiles(response.data.data);
-      } catch (error) {
-        // toast(error.message);
-      } finally {
-        setLoading(false);
-      }
-    };
-    fetchdata();
-  }, []);
-  return (
-    <Container>
-      {loading && <Spinner height="40vh" />}
-      {profiles.length === 0 && !loading && <h2>No profile found</h2>}
-      <StyledProfileList>
-        {profiles.length > 0
+const ProfileList = ({ profiles }) => (
+  <Container>
+    {profiles.length === 0 && <H2>No profiles found</H2>}
+    <StyledProfileList>
+      {profiles.length > 0
           && profiles.map((profile) => (
             <PersonPreview
               key={profile.id}
@@ -41,8 +19,19 @@ const ProfileList = () => {
               dateOfIncident={profile.date_of_incident}
             />
           ))}
-      </StyledProfileList>
-    </Container>
-  );
-};
+    </StyledProfileList>
+  </Container>
+);
+
 export default ProfileList;
+
+ProfileList.propTypes = {
+  profiles: PropTypes.arrayOf(
+    PropTypes.shape({
+      id: PropTypes.number.isRequired,
+      images: PropTypes.array.isRequired,
+      full_name: PropTypes.string.isRequired,
+      date_of_incident: PropTypes.string.isRequired
+    })
+  ).isRequired
+};
